Created a priming GPS
In the AppDelegate I have a globalinstance starting up at boot to start
giving us an initial GPS reading, this will prevent crashes of things
hopefully and allow us to start a trip without dead locations while we
focus.
Using the “Started” variable in TripData to track whether the current
instance of LocationManager is a user made trip or a priming trip.